components folder was renamed
